@@ -1,28 +1,22 @@
 using DotNetHW2;
 using Service;
-using Xunit.Sdk;
 
 namespace ServiceTest;
 
 public class ItemServiceTest
 {
-    private ItemService _sut;
-
-    public ItemServiceTest()
-    {
-        _sut = new ItemService();
-    }
+    private ItemService _sut = new();
 
     [Fact]
     public void Test_AddItem_Success()
     {
         // Arrange
-        bool expected = true;
-        
+        const bool expected = true;
+
         // Act
         ItemService.User = Admin.GetAdmin();
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        var actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
@@ -31,26 +25,26 @@ public class ItemServiceTest
     public void Test_AddItem_Fail()
     {
         // Arrange
-        bool expected = false;
-        
+        const bool expected = false;
+
         // Act
-        ItemService.User = new NonAdmin("username", "password",100);
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        ItemService.User = new NonAdmin("username", "password", 100);
+        var actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
-    
+
     [Fact]
     public void Test_DeleteItem_Success()
     {
         // Arrange
-        bool expected = true;
-        
+        const bool expected = true;
+
         // Act
         ItemService.User = Admin.GetAdmin();
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        var actual = _sut.Delete(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
@@ -59,26 +53,26 @@ public class ItemServiceTest
     public void Test_DeleteItem_Fail()
     {
         // Arrange
-        bool expected = false;
-        
+        const bool expected = false;
+
         // Act
-        ItemService.User = new NonAdmin("username", "password",100);
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        ItemService.User = new NonAdmin("username", "password", 100);
+        var actual = _sut.Delete(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
-    
+
     [Fact]
     public void Test_PurchaseItem_Success()
     {
         // Arrange
-        bool expected = false;
-        
+        const bool expected = false;
+
         // Act
         ItemService.User = Admin.GetAdmin();
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        var actual = _sut.Purchase(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
@@ -87,12 +81,12 @@ public class ItemServiceTest
     public void Test_PurchaseItem_Fail()
     {
         // Arrange
-        bool expected = true;
-        
+        const bool expected = true;
+
         // Act
-        ItemService.User = new NonAdmin("username", "password",100);
-        bool actual = _sut.Add(new GroceryItem("Grocery Item", 50, 10));
-        
+        ItemService.User = new NonAdmin("username", "password", 100);
+        var actual = _sut.Purchase(new GroceryItem("Grocery Item", 50, 10));
+
         // Assert
         Assert.Equal(expected, actual);
     }
