@@ -4,35 +4,7 @@ namespace DotNetHW2;
 
 public abstract class User
 {
-    public string? Username { get; set; }
-    public string? Password { get; set; }
+    public virtual string? Username { get; set; }
+    public virtual string? Password { get; set; }
     public static List<User> AllUsers { get; } = new List<User>();
-}
-
-public class Admin : User
-{
-    private static Admin MyAdmin { get; } = new("admin", "admin");
-
-    private Admin(string username, string password)
-    {
-        Username = username;
-        Password = password;
-    }
-
-    public static Admin GetAdmin()
-    {
-        return MyAdmin;
-    }
-}
-
-public class NonAdmin : User
-{
-    public double Money { get; set; }
-
-    public NonAdmin(string username, string password, double money)
-    {
-        Username = username;
-        Password = password;
-        Money = money;
-    }
 }
